Rename old XGBoost implementation to xgblr
