--- flcore/models/xgb/__init__.py
@@ -1,4 +1,0 @@
-import flcore.models.xgb.client
-import flcore.models.xgb.server
-import flcore.models.xgb.fed_custom_strategy
-import flcore.models.xgb.utils
