+++ flcore/models/xgblr/__init__.py
@@ -1,0 +1,4 @@
+import flcore.models.xgblr.client
+import flcore.models.xgblr.server
+import flcore.models.xgblr.fed_custom_strategy
+import flcore.models.xgblr.utils
