+++ flcore/server_selector.py
@@ -1,6 +1,6 @@
 #import flcore.models.logistic_regression.server as logistic_regression_server
 #import flcore.models.logistic_regression.server as logistic_regression_server
-import flcore.models.xgb.server as xgb_server
+import flcore.models.xgblr.server as xgblr_server
 import flcore.models.random_forest.server as random_forest_server
 import flcore.models.linear_models.server as linear_models_server
 import flcore.models.weighted_random_forest.server as weighted_random_forest_server
@@ -22,8 +22,8 @@ def get_model_server_and_strategy(config, data=None):
             config
         )
         
-    elif model == "xgb":
-        server, strategy = xgb_server.get_server_and_strategy(config, data)
+    elif model == "xgblr":
+        server, strategy = xgblr_server.get_server_and_strategy(config, data)
     else:
         raise ValueError(f"Unknown model: {model}")
